+++ platform/cc/Bitmap.cc
@@ -190,13 +190,27 @@ extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nNotify
     instance->notifyPixelsChanged();
 }
 
-extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nEraseColor
+extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nErase4f
+  (JNIEnv* env, jclass jclass, jlong ptr, jfloat r, jfloat g, jfloat b, jfloat a, jlong colorSpacePtr) {
+    SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
+    SkColorSpace* colorSpace = reinterpret_cast<SkColorSpace*>(static_cast<uintptr_t>(colorSpacePtr));
+    instance->eraseColor({r, g, b, a}, colorSpace);
+}
+
+extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nEraseRect4f
+  (JNIEnv* env, jclass jclass, jlong ptr, jfloat r, jfloat g, jfloat b, jfloat a, jlong colorSpacePtr, jint left, jint top, jint right, jint bottom) {
+    SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
+    SkColorSpace* colorSpace = reinterpret_cast<SkColorSpace*>(static_cast<uintptr_t>(colorSpacePtr));
+    instance->erase({r, g, b, a}, colorSpace, {left, top, right, bottom});
+}
+
+extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nErase
   (JNIEnv* env, jclass jclass, jlong ptr, jint color) {
     SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
     instance->eraseColor(color);
 }
 
-extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nErase
+extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_Bitmap__1nEraseRect
   (JNIEnv* env, jclass jclass, jlong ptr, jint color, jint left, jint top, jint right, jint bottom) {
     SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
     instance->erase(color, {left, top, right, bottom});
@@ -206,6 +220,13 @@ extern "C" JNIEXPORT jint JNICALL Java_io_github_humbleui_skija_Bitmap__1nGetCol
   (JNIEnv* env, jclass jclass, jlong ptr, jint x, jint y) {
     SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
     return instance->getColor(x, y);
+}
+
+extern "C" JNIEXPORT jobject JNICALL Java_io_github_humbleui_skija_Bitmap__1nGetColor4f
+  (JNIEnv* env, jclass jclass, jlong ptr, jint x, jint y) {
+    SkBitmap* instance = reinterpret_cast<SkBitmap*>(static_cast<uintptr_t>(ptr));
+    SkColor4f color = instance->getColor4f(x, y);
+    return env->NewObject(skija::Color4f::cls, skija::Color4f::ctor, color.fR, color.fG, color.fB, color.fA);
 }
 
 extern "C" JNIEXPORT jfloat JNICALL Java_io_github_humbleui_skija_Bitmap__1nGetAlphaf