+++ shared/java/Bitmap.java
@@ -695,12 +695,96 @@ public class Bitmap extends Managed implements IHasImageInfo {
     }
 
     /**
+     * Replaces pixel values with color, interpreted as being in the colorSpace.
+     * All pixels contained by getBounds() are affected. If the getColorType() is
+     * {@link ColorType#GRAY_8} or {@link ColorType#RGB_565}, then alpha is ignored; RGB is
+     * treated as opaque. If getColorType() is {@link ColorType#ALPHA_8}, then RGB is ignored.
+     *
+     * @param color      unpremultiplied color
+     * @param colorSpace SkColorSpace describing the encoding of color
+     * @return           this
+     *
+     * @see <a href="https://fiddle.skia.org/c/@Bitmap_eraseColor">https://fiddle.skia.org/c/@Bitmap_eraseColor</a>
+     */
+    @NotNull @Contract("_, _ -> this")
+    public Bitmap erase(@NotNull Color4f color, @Nullable ColorSpace colorSpace) {
+        try {
+            assert color != null : "Bitmap::erase expected color != null";
+            Stats.onNativeCall();
+            _nErase4f(_ptr, color._r, color._g, color._b, color._a, Native.getPtr(colorSpace));
+            return this;
+        } finally {
+            Reference.reachabilityFence(this);
+        }
+    }
+
+    /**
+     * Replaces pixel values with color.
+     * All pixels contained by getBounds() are affected. If the getColorType() is
+     * {@link ColorType#GRAY_8} or {@link ColorType#RGB_565}, then alpha is ignored; RGB is
+     * treated as opaque. If getColorType() is {@link ColorType#ALPHA_8}, then RGB is ignored.
+     *
+     * @param color  unpremultiplied color
+     * @return       this
+     *
+     * @see <a href="https://fiddle.skia.org/c/@Bitmap_eraseColor">https://fiddle.skia.org/c/@Bitmap_eraseColor</a>
+     */
+    @NotNull @Contract("_ -> this")
+    public Bitmap erase(@NotNull Color4f color) {
+        return erase(color, (ColorSpace) null);
+    }
+
+    /**
+     * Replaces pixel values with color, interpreted as being in the colorSpace.
+     * All pixels contained by getBounds() are affected. If the getColorType() is
+     * {@link ColorType#GRAY_8} or {@link ColorType#RGB_565}, then alpha is ignored; RGB is
+     * treated as opaque. If getColorType() is {@link ColorType#ALPHA_8}, then RGB is ignored.
+     *
+     * @param color      unpremultiplied color
+     * @param colorSpace SkColorSpace describing the encoding of color
+     * @param area       rectangle to fill
+     * @return           this
+     *
+     * @see <a href="https://fiddle.skia.org/c/@Bitmap_erase">https://fiddle.skia.org/c/@Bitmap_erase</a>
+     */
+    @NotNull @Contract("_, _, _ -> this")
+    public Bitmap erase(@NotNull Color4f color, @Nullable ColorSpace colorSpace, @NotNull IRect area) {
+        try {
+            assert color != null : "Bitmap::erase expected color != null";
+            assert area != null : "Bitmap::erase expected area != null";
+            Stats.onNativeCall();
+            _nEraseRect4f(_ptr, color._r, color._g, color._b, color._a, Native.getPtr(colorSpace), area._left, area._top, area._right, area._bottom);
+            return this;
+        } finally {
+            Reference.reachabilityFence(this);
+        }
+    }
+
+    /**
+     * Replaces pixel values with color.
+     * All pixels contained by getBounds() are affected. If the getColorType() is
+     * {@link ColorType#GRAY_8} or {@link ColorType#RGB_565}, then alpha is ignored; RGB is
+     * treated as opaque. If getColorType() is {@link ColorType#ALPHA_8}, then RGB is ignored.
+     *
+     * @param color  unpremultiplied color
+     * @param area   rectangle to fill
+     * @return       this
+     *
+     * @see <a href="https://fiddle.skia.org/c/@Bitmap_eraseColor">https://fiddle.skia.org/c/@Bitmap_eraseColor</a>
+     */
+    @NotNull @Contract("_, _ -> this")
+    public Bitmap erase(@NotNull Color4f color, @NotNull IRect area) {
+        return erase(color, null, area);
+    }
+
+    /**
      * Replaces pixel values with color, interpreted as being in the sRGB ColorSpace.
      * All pixels contained by getBounds() are affected. If the getColorType() is
      * {@link ColorType#GRAY_8} or {@link ColorType#RGB_565}, then alpha is ignored; RGB is
      * treated as opaque. If getColorType() is {@link ColorType#ALPHA_8}, then RGB is ignored.
      *
      * @param color  unpremultiplied color
+     * @return           this
      *
      * @see <a href="https://fiddle.skia.org/c/@Bitmap_eraseColor">https://fiddle.skia.org/c/@Bitmap_eraseColor</a>
      */
@@ -708,7 +792,7 @@ public class Bitmap extends Managed implements IHasImageInfo {
     public Bitmap erase(int color) {
         try {
             Stats.onNativeCall();
-            _nEraseColor(_ptr, color);
+            _nErase(_ptr, color);
             return this;
         } finally {
             Reference.reachabilityFence(this);
@@ -725,6 +809,7 @@ public class Bitmap extends Managed implements IHasImageInfo {
      *
      * @param color  unpremultiplied color
      * @param area   rectangle to fill
+     * @return           this
      *
      * @see <a href="https://fiddle.skia.org/c/@Bitmap_erase">https://fiddle.skia.org/c/@Bitmap_erase</a>
      */
@@ -732,7 +817,7 @@ public class Bitmap extends Managed implements IHasImageInfo {
     public Bitmap erase(int color, @NotNull IRect area) {
         try {
             Stats.onNativeCall();
-            _nErase(_ptr, color, area._left, area._top, area._right, area._bottom);
+            _nEraseRect(_ptr, color, area._left, area._top, area._right, area._bottom);
             return this;
         } finally {
             Reference.reachabilityFence(this);
@@ -759,6 +844,31 @@ public class Bitmap extends Managed implements IHasImageInfo {
         try {
             Stats.onNativeCall();
             return _nGetColor(_ptr, x, y);
+        } finally {
+            Reference.reachabilityFence(this);
+        }
+    }
+
+    /** 
+     * <p>Returns pixel at (x, y) as unpremultiplied color.
+     * Returns black with alpha if ColorType is {@link ColorType#ALPHA_8}.</p>
+     *
+     * <p>Input is not validated: out of bounds values of x or y returns undefined values
+     * or may crash if. Fails if ColorType is {@link ColorType#UNKNOWN} or
+     * pixel address is nullptr.</p>
+     *
+     * <p>ColorSpace in ImageInfo is ignored. Some color precision may be lost in the
+     * conversion to unpremultiplied color; original pixel data may have additional
+     * precision.</p>
+     *
+     * @param x  column index, zero or greater, and less than getWidth()
+     * @param y  row index, zero or greater, and less than getHeight()
+     * @return   pixel converted to unpremultiplied color
+     */
+    public Color4f getColor4f(int x, int y) {
+        try {
+            Stats.onNativeCall();
+            return _nGetColor4f(_ptr, x, y);
         } finally {
             Reference.reachabilityFence(this);
         }
@@ -998,9 +1108,12 @@ public class Bitmap extends Managed implements IHasImageInfo {
     @ApiStatus.Internal public static native boolean _nIsReadyToDraw(long ptr);
     @ApiStatus.Internal public static native int     _nGetGenerationId(long ptr);
     @ApiStatus.Internal public static native void    _nNotifyPixelsChanged(long ptr);
-    @ApiStatus.Internal public static native void    _nEraseColor(long ptr, int color);
-    @ApiStatus.Internal public static native void    _nErase(long ptr, int color, int left, int top, int right, int bottom);
+    @ApiStatus.Internal public static native void    _nErase4f(long ptr, float r, float g, float b, float a, long colorSpacePtr);
+    @ApiStatus.Internal public static native void    _nEraseRect4f(long ptr, float r, float g, float b, float a, long colorSpacePtr, int left, int top, int right, int bottom);
+    @ApiStatus.Internal public static native void    _nErase(long ptr, int color);
+    @ApiStatus.Internal public static native void    _nEraseRect(long ptr, int color, int left, int top, int right, int bottom);
     @ApiStatus.Internal public static native int     _nGetColor(long ptr, int x, int y);
+    @ApiStatus.Internal public static native Color4f _nGetColor4f(long ptr, int x, int y);
     @ApiStatus.Internal public static native float   _nGetAlphaf(long ptr, int x, int y);
     @ApiStatus.Internal public static native boolean _nExtractSubset(long ptr, long dstPtr, int left, int top, int right, int bottom);
     @ApiStatus.Internal public static native byte[]  _nReadPixels(long ptr, int width, int height, int colorType, int alphaType, long colorSpacePtr, long dstRowBytes, int srcX, int srcY);